Adapt NixFreezeCommand.process_requirements
Adjusted the code, so that it mainly relies on the pip models and not on the
internals of the resolver anymore.

A change in the resolver in pip is that it does now return a RequirementSet and
not modify things in place anymore. The input to resolve() is now a list of
InstallRequirement.